refactor(calculator): 패키지 분리
