--- src/main/java/camp/nextstep/edu/calculator/Expression.java
@@ -1,8 +1,0 @@
-package camp.nextstep.edu.calculator;
-
-interface Expression {
-
-    Number sumAll();
-
-    boolean contains(Expression expression);
-}
