+++ src/main/java/camp/nextstep/edu/calculator/Guard.java
@@ -2,12 +2,12 @@ package camp.nextstep.edu.calculator;
 
 import java.util.Objects;
 
-final class Guard {
+public final class Guard {
 
     private Guard() {
     }
 
-    static boolean isNullOrBlank(final String value) {
+    public static boolean isNullOrBlank(final String value) {
         return Objects.isNull(value) || value.isBlank();
     }
 }
