+++ src/main/java/camp/nextstep/edu/calculator/expression/Expression.java
@@ -1,0 +1,10 @@
+package camp.nextstep.edu.calculator.expression;
+
+import camp.nextstep.edu.calculator.value.Number;
+
+public interface Expression {
+
+    Number sumAll();
+
+    boolean contains(Expression expression);
+}
