+++ src/main/java/camp/nextstep/edu/calculator/value/Number.java
@@ -1,4 +1,4 @@
-package camp.nextstep.edu.calculator;
+package camp.nextstep.edu.calculator.value;
 
 public interface Number {
