+++ src/main/java/camp/nextstep/edu/calculator/value/Positive.java
@@ -1,4 +1,7 @@
-package camp.nextstep.edu.calculator;
+package camp.nextstep.edu.calculator.value;
+
+import camp.nextstep.edu.calculator.Guard;
+import camp.nextstep.edu.calculator.InvalidArgumentException;
 
 import java.util.Map;
 import java.util.Objects;
