+++ src/main/java/camp/nextstep/edu/calculator/value/Value.java
@@ -1,0 +1,17 @@
+package camp.nextstep.edu.calculator.value;
+
+import camp.nextstep.edu.calculator.Guard;
+
+public final class Value {
+
+    private Value() {
+    }
+
+    public static Number of(final String source) {
+        if (Guard.isNullOrBlank(source) || Zero.INSTANCE.equalsBy(source)) {
+            return Zero.INSTANCE;
+        }
+
+        return Positive.of(source);
+    }
+}
