+++ src/test/java/camp/nextstep/edu/calculator/expression/StringExpressionTest.java
@@ -1,5 +1,7 @@
-package camp.nextstep.edu.calculator;
+package camp.nextstep.edu.calculator.expression;
 
+import camp.nextstep.edu.calculator.value.Number;
+import camp.nextstep.edu.calculator.value.Value;
 import org.assertj.core.api.AbstractBooleanAssert;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
